+++ src/main/resources/static/css/user/diary/list.css
@@ -1,21 +1,16 @@
 
-/* html,body{
-    width: 100%;
-    height: 100%;
-} */
+
 #main{
-    /* width: 100%;
-    height: 100%; */
-    width: 100%;
+    width: 1080px;
     height: 600px;
     /* min-height: 600px; */
 }
 
-#main .list{
+#main .book .page2 .list{
     display: flex;
     flex-wrap: wrap;
 }
-#main .list .panel {
+#main .book .page2 .list .panel {
     display: flex;
     width: 450px;
     height: 314px;
@@ -29,11 +24,11 @@
 }
 
 
-#main .list .panel .diary{
+#main .book .page2 .list .panel .diary{
     border: 1px solid #ccc;
 }
 
-    #main .list .panel .diary .head{
+#main .book .page2 .list .panel .diary .head{
     display: flex;
     width: 100%;
     height: 64px;
@@ -45,14 +40,14 @@
     border: 1px 1px solid #ccc;
 }
 
-    #main .list .panel .diary .head .deco{
+#main .book .page2 .list .panel .diary .head .deco{
     
     width: 64px;
     height: 64px;
     /* background: khaki; */  
 }
 
-    #main .list .panel .diary .head .deco .reborn{
+#main .book .page2 .list .panel .diary .head .deco .reborn{
     display: inline-block;
     width: 99px;
     height: 24px;
@@ -65,7 +60,7 @@
 }
 
 
-    #main .list .panel .diary .head .title{
+#main .book .page2 .list .panel .diary .head .title{
     display: inline-block;
     color: #4a4a4a;
     text-align: center;
@@ -75,14 +70,14 @@
     margin-right: 64px;    
 }
 
-    #main .list .panel .diary .body{
+#main .book .page2 .list .panel .diary .body{
     /* background: red; */
     width: 100%;
     height: 250px;
     
 }
 
-    #main .list .diary .body .content{
+#main .book .page2 .list .diary .body .content{
     display: inline-block;
     padding:15px;
     
@@ -102,7 +97,7 @@
     width: 100%; height: 100%;
     text-align: center;
     box-shadow: 5px 5px 15px #333;
-    background: blue;
+    /* background: blue; */
 }
 
 .book .page1{
@@ -121,7 +116,10 @@
     box-sizing: border-box;
 }
 .book .page2{
-    background: #cba;
+    /* background: #cba; */
+    background-image: url(../../../images/diary/paper2.jpg);
+    background-repeat: no-repeat;
+    background-size: cover;
     trnasform-origin: left center;
     transition: 2s;   
     /* z-index:2; */
