+++ src/main/resources/static/user/diary/edit.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Document</title>
+        <link rel="stylesheet"  href="../../css/reset.css">
+        <link rel="stylesheet" href="../../css/index.css">
+        <link rel="stylesheet" href="../../css/user/diary/visual.css">
+    
+        <link rel="stylesheet" href="../../css/user/diary/reg.css">
+        <link rel="preconnect" href="https://fonts.gstatic.com">
+        <link href="https://fonts.googleapis.com/css2?family=Poor+Story&display=swap" rel="stylesheet">
+    </head>
+    <body>
+        <!-----------------------------------------------header------------------------------------->
+        <header id="header" class="header">
+            <div class="sub-header">
+                <a href="/index.html"><img class="logo" src="../../images/logo.png"></a>
+                <nav>
+                    <ul>
+                        <li><a href="">커플노트</a></li>
+                        <li><a href="">커뮤니티</a></li>
+                        <li><a href="">로그인</a></li>
+                    </ul>
+                </nav>
+            </div>
+        </header>
+        <!----------------------------------------------visual------------------------------------->
+        <section id="visual">
+            Diary
+        </section>
+        <!----------------------------------------------main--------------------------------------->
+    
+    <section id="main">
+        <div class="panel">
+            <div class="well">
+                <span>일기장 글 수정하기</span>
+            </div>
+            <form name="reg" action="" method="POST" class="form">
+                <div class="title">
+                    <label for="input-title" >제목  </label>
+                    <input type="text" name="title" id="input-title">
+                </div>
+                <span class="mood-span">기분</span>
+                <div class="mood">
+                    <input type="radio" name="mood" value="smile1" id="smile1">
+                    <label for="smile1"><img src="../../images/diary/smile1.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="smile2" id="smile2">
+                    <label for="smile2"><img src="../../images/diary/smile2.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="smile3" id="smile3">
+                    <label for="smile3"><img src="../../images/diary/smile3.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="normal" id="normal">
+                    <label for="normal"><img src="../../images/diary/normal.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="love" id="love">
+                    <label for="love"><img src="../../images/diary/love.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="sad" id="sad">
+                    <label for="sad"><img src="../../images/diary/sad.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="upset" id="upset">
+                    <label for="upset"><img src="../../images/diary/upset.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="angry" id="angry">
+                    <label for="angry"><img src="../../images/diary/angry.png" width="50px" height="50px"></label>
+                    <input type="radio" name="mood" value="sunglasses" id="sunglasses">
+                    <label for="sunglasses"><img src="../../images/diary/sunglasses.png" width="50px" height="50px"></label>
+                </div>
+                
+                <div class="content">
+                    <label for="input-content" >내용  </label>
+                    <textarea name="" id="" cols="30" rows="10"></textarea>
+                </div>
+                <div class="btns">
+                    <input type="submit" value="수정" class="button">
+                    <a href="http://www.naver.com" class="button">취소</a>
+                </div>
+            </form>
+        </div>
+    </section>
+    <!---------------------------------------footer----------------------------------------------->
+    <footer id="footer" class="footer">
+        <div>
+            <ul>
+                <li>
+                    <a href="">회사소개</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li>
+                    <a href="">이용약관</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li>
+                    <a href="">개인정보처리방침</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li>
+                    <a href="">사업자정보확인</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li>
+                    <a href="">틈 마케팅센터</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li>
+                    <a href="">액티비티 호스트센터</a>
+                    <span>&nbsp;&nbsp;|&nbsp;&nbsp;</span>
+                </li>
+                <li><a href="">콘텐츠산업진흥법에의한 표시</a></li>
+            </ul>
+
+            <p><b>고객행복센터 1577-8282</b> 오전 09:00 ~ 오후18:00</p>
+
+            <address>
+                <span>(주)또바기</span>
+                주소 : 서울특별시 종로구 인사동길 12 대일빌딩 15층
+                <br />대표이사 : 김병준 | 사업자등록번호 777-11-45678 <br />통신판매번호 : 2020-서울강남-12345 | 관광사업자 등록번호 : 제
+                1324-55호 <br />전화번호 : 010-1111-2222 <br />전자우편주소 : ttobagi@ttobagi.co.kr <br />Copyright ttobagi COMPANY Corp. All rights
+                reserved.
+            </address>
+        </div>
+    </footer>
+</body>
+</html>
