+++ src/main/resources/static/user/diary/list.html
@@ -61,7 +61,10 @@
                     </div>
                 </div>
             </div>
+            
+
         </div>
+        <button class="button" >열기</button>
     </section>
     <!-- <section id="book">
         <div class="page">
