+++ src/main/resources/static/user/diary/reg.html
@@ -69,7 +69,7 @@
                     <textarea name="" id="" cols="30" rows="10"></textarea>
                 </div>
                 <div class="btns">
-                    <input type="submit" value="보내기" class="button">
+                    <input type="submit" value="쓰기" class="button">
                     <a href="http://www.naver.com" class="button">취소</a>
                 </div>
             </form>
